Added ft_build_branch_utils2.c
