+++ includes/tokenize.h
@@ -130,12 +130,13 @@ void	*ft_build_tree(t_list **token_list, t_node **parent_node);
 void	*ft_build_branch(t_list **list, t_exec *exec, t_node *sub_root);
 //ft_built_branch_utils.c
 t_list	**ft_get_args(t_list **list);
-bool	ft_find_next_redir(t_list **list);
-t_list	**ft_create_sub_list(t_list **list);
-t_node	*ft_create_subroot_node(t_list **list);
 t_exec	*ft_create_exec_node(t_token *token, t_list **list);
+bool	ft_find_next_redir(t_list **list);
 void	ft_assign_redir_mode(t_redir **redir);
 t_redir	*ft_init_redir(t_token *token, t_list **list);
+//ft_built_branch_utils2.c
+t_list	**ft_create_sub_list(t_list **list);
+t_node	*ft_create_subroot_node(t_list **list);
 
 //ft_free_tree.c
 void	ft_free_tree(void *root);
